CompoundParser: real numbers now supported for subscripts
In order to enable this, nAtomsAll from the CompoundData struct was
changed to double from int

All bindings and examples updated

+++ include/xraylib-parser.h
@@ -50,7 +50,7 @@ THIS SOFTWARE IS PROVIDED BY Tom Schoonjans ''AS IS'' AND ANY EXPRESS OR IMPLIED
 
 struct compoundData {
 	int nElements;
-	int nAtomsAll;
+	double nAtomsAll;
 	int *Elements;
 	double *massFractions;
 };
